+++ usr/src/uts/sun4u/sys/zulumod.h
@@ -2,9 +2,8 @@
  * CDDL HEADER START
  *
  * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
+ * Common Development and Distribution License (the "License").
+ * You may not use this file except in compliance with the License.
  *
  * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
  * or http://www.opensolaris.org/os/licensing.
@@ -20,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2006 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -125,7 +124,7 @@ typedef struct {
 	struct zuluvm_proc *proc1;
 	struct zuluvm_proc *proc2;
 	volatile uint32_t state;	/* state of tlb miss handling */
-	short		intr_num;	/* our soft intr number */
+	uint64_t	intr_num;	/* our soft intr number */
 	short		dmv_intr;	/* dmv interrupt handle */
 #ifdef ZULUVM_STATS
 	int 		cancel;
